+++ src/TokensTracker/TokensTracker.cs
@@ -3,6 +3,7 @@ using Neo.IO;
 using Neo.Ledger;
 using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
+using Neo.Plugins.Trackers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -40,9 +41,9 @@ namespace Neo.Plugins
             string path = string.Format(_dbPath, neoSystem.Settings.Network.ToString("X8"));
             _db = neoSystem.LoadStore(GetFullPath(path));
             if (_enabledTrackers.Contains("NEP-11"))
-                trackers.Add(new Nep11Tracker(_db, _maxResults, _shouldTrackHistory, neoSystem));
+                trackers.Add(new Trackers.NEP_11.Nep11Tracker(_db, _maxResults, _shouldTrackHistory, neoSystem));
             if (_enabledTrackers.Contains("NEP-17"))
-                trackers.Add(new Nep17Tracker(_db, _maxResults, _shouldTrackHistory, neoSystem));
+                trackers.Add(new Trackers.NEP_17.Nep17Tracker(_db, _maxResults, _shouldTrackHistory, neoSystem));
             foreach (TrackerBase tracker in trackers)
                 RpcServerPlugin.RegisterMethods(tracker, _network);
         }
