+++ src/TokensTracker/Trackers/NEP-17/Nep17TransferKey.cs
@@ -1,13 +1,7 @@
 using Neo.IO;
 using System;
-using System.Buffers.Binary;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Neo.Plugins.Storage
+namespace Neo.Plugins.Trackers.NEP_17
 {
     public class Nep17TransferKey : TokenTransferKey, IComparable<Nep17TransferKey>, IEquatable<Nep17TransferKey>, ISerializable
     {
