+++ src/TokensTracker/Trackers/TokenBalance.cs
@@ -1,13 +1,8 @@
 using Neo.IO;
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Numerics;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Neo.Plugins.Storage
+namespace Neo.Plugins.Trackers
 {
     public class TokenBalance : ISerializable
     {
